Assistant checkpoint: Update getTickPrices method to use correct API format
Assistant generated file changes:
- client/src/lib/xtb-service.ts: Update getTickPrices to use correct subscription format

---

User prompt:

Adjust the code to call for latest tick price for USDBRL using this instead:

Command: getTickPrices
Description: Establishes subscription for quotations and allows to obtain the relevant information in real-time, as soon as it is available in the system. The getTickPrices  command can be invoked many times for the same symbol, but only one subscription for a given symbol will be created. Please beware that when multiple records are available, the order in which they are received is not guaranteed.

Subscribe format:
Parameters:

name	type	description
symbol	String	Symbol
minArrivalTime	Number	This field is optional and defines the minimal interval in milliseconds between any two consecutive updates. If this field is not present, or it is set to 0 (zero), ticks - if available - are sent to the client with interval equal to 200 milliseconds. In order to obtain ticks as frequently as server allows you, set it to 1 (one).
maxLevel	Number	This field is optional and specifies the maximum level of the quote that the user is interested in. If this field is not specified, the subscription is active for all levels that are managed in the system.
Example:

{
	"command": "getTickPrices",
	"streamSessionId": "8469308861804289383",
	"symbol": "EURUSD",
	"minArrivalTime": 5000,
	"maxLevel": 2
}